lib/rb: Namespace properly the std libs
