+++ lib/rb/lib/thrift/types/known/duration/duration_types.rb
